requirements.txt
fastapi
uvicorn
google-generativeai
pymongo
bcrypt
python-jose[cryptography]
passlib[bcrypt]